[WMCB] switch from winrm to SSH
this commit makes the switch from winRM to SSH in wmcb tests. Also
removes winRM altogether. SSH is now default to run any command on VM.

Also, updated the powershell script to work with SSH. This was needed as
the current wget powershell script was throwing an error and not
downloading the files as required.